library: rci: provide access to configuration
This commit implements rci session start and end handlers so that any RCI
handler can interact with the connector configuration. This allows:
 * Read access for RCI getters.
 * Read and write access for RCI setters.
RCI handlers may interact with the connector configuration using the cc_cfg
struct.

The configuration is read (and saved, if needed) to the configuration file once
per RCI call.

https://jira.digi.com/browse/DEL-3287

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>