library: config: allow Vendor IDs bigger than 0xFE000000
We need to establish a maximum of 0xFFFFFFFF for the Vendor ID to be able to
use value 0xFE080003 (Embedded Linux Vendor ID).

The entry in the configuration file must be read as a string and the transformed
to a numeric value, to avoid the 'libconfuse' limitation when reading numeric
values (signed longs).

https://jira.digi.com/browse/DEL-3566

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>