library: system monitor: adjust the rate and samples to store in every loop
The monitor system settings can be modified in runtime, this includes the sample
rate, the number of samples to store per parameters, and the parameters to
monitor. Under these circumstances we need to determine the total number of
samples to store (of all enabled parameters) and the sample rate in every loop
of the system monitor thread.

https://jira.digi.com/browse/DEL-3596
https://jira.digi.com/browse/DEL-3597
https://jira.digi.com/browse/DEL-3598

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>