library: rci: static location: do not allow empty or invalid values (NaN)
When setting a parameter, an empty value from the
'Configuration > Static Location' page of Device Cloud is translated to a NaN
value and sent to the device (seems it only happens with float parameters, not
with integers).

If NaN is received to be configured for 'latitude', 'longitude', and 'altitude',
a bad value error is returned.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>