rci_state_device_info: fix string copies
The current code uses the following call several times:

strncpy(a, b, strlen(b));

This has two big problems:
* It does not copy the null terminator from b into a. The last parameter
  should really be the size available in a, not the size of b.
* Even if we change the last parameter to the size of a, strncpy still
  has a big problem: if the dest string is longer than the size, it won't
  copy the null terminator.

Most of the strncpy calls can just be removed by using a different initial
value of the string. For the few cases in which it cannot be omitted, use
snprintf instead which solves the problems.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>