device_info: simplify function to read output of commands
The function had a loop to read through lines and concatenate strings but
it was working wrong and didn't initialize the 'out' array before doing
the first concatenation.

This funcion is only used to get the output of 'uname -a' but we need it
for another command to get the MCA address. In both cases, such commands
only return one line. This commit simplifies the function to return that
one line only and also removes the new line char produced by command
outputs.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>