app: remove atexit to avoid segmentation fault when sending signal SIGINT
While the process is running if it receives a SIGINT signal a segmentation
fault is thrown. It seems it is caused because an 'atexit' function is
registered in the application while another is registered for openssl
('atexit(OPENSSL_cleanup)')

This commit removes the 'atexit' from the application and reworks the use of
signals.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>