app: monitor user button and send data point
The alias USER_BUTTON is used for the button. It must be defined inside file
'/etc/libdigiapix.conf'.

The application listens for rising and falling edges on the USER_BUTTON and
uploads those values together. That is, pressed and released button events are
sent together when the button is released.

https://onedigi.atlassian.net/browse/DEL-7899

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>