library: system_monitor: fix segmentation fault on 32-bits systems
Fix a segmentation fault error on 32-bits systems while creating datapoints for the
new network interfaces measures.

The value of the 'state' datapoints was being passed as 'unsigned long long' (8 bytes)
to the argv list during the datapoint creation. However, this parameter was being defined
and read as an 'int32_t' in the datapoint stream. The cloud connector defines the 'int32_t'
type as a 'long int', which results in a size of 4 bytes in 32-bits systems and 8 bytes
in 64-bit systems. In 64-bit systems the data size passed and read was always 8 bytes so
that the problem was not visible. In 32-bits system, the difference of 4 bytes in the data
size passed and the data size read was causing a memory padding of 4 bytes. The result was
undefined memory reads in subsequent datapoint creation calls, ending in a segmentation fault.

The fix is to configure the 'state' stream datapoint as an 'int64_t' so that the size of
the value passed and read in the datapoint creation function always matches in 32 and 64-bits
systems (8 bytes).

Signed-off-by: David Escalona <david.escalona@digi.com>
Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>